requirements.txt
httpcore==0.16.3
httpx==0.23.3
nb-cli==1.0.5
nonebot-adapter-onebot==2.2.1
nonebot2==2.0.0rc3
Pillow==9.4.0
virtualenv==20.17.1
bilibili-api-python